+++ mobile/package.json
@@ -7,8 +7,12 @@
 		"test": "jest"
 	},
 	"dependencies": {
-		"react": "16.0.0-alpha.12",
-		"react-native": "0.46.4"
+		"react": "16.0.0-beta.1",
+		"react-native": "0.46.4",
+		"react-router-native": "4.1.1",
+
+	    "apollo-client": "1.9.0-1",
+	    "react-apollo": "^1.4.8"
 	},
 	"devDependencies": {
 		"babel-jest": "20.0.3",
